prog petit robot setlowspeedForward et backward changement de code pour la detection et les collisions suppression des ignores detection

+++ src/Driver-EV3/LegoIRSensor.hpp
@@ -46,7 +46,7 @@ public:
     /*!
      * \brief donne la distance detectée en mm.
      */
-    double getDistanceMM();
+    float getDistanceMM();
 
 };
